requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coverage==4.5.3
Flask==1.1.1
flask-csrf==0.9.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
nose2==0.9.1
pycrypto==2.6.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
